fix color action buttons
